add index.html with some docs
