+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon -r  esm serverAPI.js"
+    "start": "nodemon -r  esm src/serverAPI.js"
   },
   "keywords": [],
   "author": "",
