+++ src/controllers/authentification/index.js
@@ -1,0 +1,3 @@
+export { default as loginController } from "./login.js";
+export { default as logoutController } from "./logout.js";
+export { refreshAuth, refreshToken } from "./refresh.js";
