+++ src/controllers/dashboard/index.js
@@ -1,0 +1,1 @@
+export { getDeviceDVBC, getDeviceStatus, getDevicePing } from "./dashboard";
