+++ src/controllers/index.js
@@ -1,0 +1,4 @@
+export { loginController, logoutController, refreshAuth, refreshToken } from "./authentification";
+export { getAllUsers, getOneUser, addUser, updateOneUser, changeUserPassword, deleteOneUser } from "./users";
+export { getProfile, updateProfile, changeProfilePassword, deleteProfile } from "./profile";
+export { getDeviceDVBC, getDeviceStatus, getDevicePing } from "./dashboard";
