+++ src/controllers/profile/index.js
@@ -1,0 +1,7 @@
+export { getProfile } from "./profile.js";
+export { updateProfile } from "./profile.js";
+export { changeProfilePassword } from "./profile.js";
+export { deleteProfile } from "./profile.js";
+
+
+
