+++ src/controllers/users/index.js
@@ -1,0 +1,7 @@
+export { getAllUsers } from "./users.js";
+export { getOneUser } from "./users.js";
+export { addUser } from "./users.js";
+export { updateOneUser } from "./users.js";
+export { changeUserPassword } from "./users.js";
+export { deleteOneUser } from "./users.js";
+
