+++ src/routs/index.js
@@ -1,5 +1,6 @@
 export { default as login } from "./login.js";
 export { default as logout } from "./logout.js";
 export { default as users } from "./users.js";
-export { default as status } from "./status.js";
+export { default as dashboard } from "./dashboard.js";
 export { default as refresh } from "./refresh.js";
+export { default as profile } from "./profile.js";
